Updated to use Generic dataset
Please note that in the original MPII dataset, the labels (annotations)
had the is_train tag, which would identify if the image is to be used
for training or validation.
Initializing Dataset would refer to the same dataset
and create the same two sets based on the is_train flag.
For YogiDB, we are creating the train/validation sets on the fly. So
We have to init it once and then set the flag for each individual
dataset.